updated author tests

+++ xt/author/05_kwalitee.t
@@ -4,4 +4,6 @@ use warnings;
 use FindBin '$Bin';
 use Test::Kwalitee;
 my $file = "$Bin/../../Debian_CPANTS.txt";
-unlink $file or die "can't unlink $file: $!\n";
+if (-e $file) {
+    unlink $file or die "can't unlink $file: $!\n";
+}
